+++ expression.cpp
@@ -593,7 +593,7 @@ DataTypeStruct ExpressionC::analyzeColumns(vector<string>* fieldnames, vector<Da
           //m_expType = UNKNOWN;
           m_datatype.datatype = UNKNOWN;
         }
-        trace(DEBUG, "Expression '%s' type is %s, data type is UNKNOWN\n", m_expStr.c_str(), decodeExptype(m_expType).c_str());
+        //trace(DEBUG, "Expression '%s' type is %s, data type is UNKNOWN\n", m_expStr.c_str(), decodeExptype(m_expType).c_str());
         return m_datatype;
       }
       if (m_datatype.datatype == UNKNOWN){